requirements.txt
streamlit>=1.32.0
langchain>=0.3.23
langchain-community>=0.3.21
langchain-google-genai>=2.0.10
faiss-cpu>=1.10.0
pypdf>=5.4.0
google-generativeai>=0.8.4
sentence-transformers>=4.1.0
python-docx>=1.1.2
ipywidgets>=8.1.6
python-dotenv>=1.1.0